Add TTLCache for get_sr_rank
Users can be in more than one guild, and when they leave OW we get
an update for each guild they are in. To avoid duplicate requests,
it's easiest to just cache the SR for a while; in the future, there
might be more cases where the same BattleTag is requested in parallel.